Detect the presence of python3-cffi module
Use the 'module' argument of find_installation() to check whether Python's CFFI
module is installed. If it is then build the shared library that is needed by
collect_stats.py, which is a script used to collect system metrics for AI model
training; if a Python installation with CFFI module is not found, then don't
bother building the shared library. (Note: collect_stats.py is generally not
needed on production/deployment)

A major behind this is that there are SLES distributions that does not provide
the python3-cffi package, which makes RPM packaging difficult. With building of
the shared library (requires python3-cffi) determined dynmically, we can get
away with the missing python3-cffi package.