Breakup integration test into smaller smaller pieces
This is a continuation of work done in #11. While building of _phoebe.abi3.so
can be toggled and dynamically determined, the integration test is not aware of
that; this works right now because testing of collect_stat.py is disabled right
now due to #2.

Break up what was run_tests.sh into setup.sh (majority of run_tests.sh goes
there) and two test() defined in /tests/meson.build, one for phoebe and another
for collect_static.py. Since meson runs the tests in parallel by default,
setup.sh is given an arbitrary high value with is_parallel == false to ensure it
is ran before anything else.