Always expose the abi target in /scripts/meson.build
The changes introduced in 2e0aef0407 to build the shared library needed by
collect_stats.py only in the presence of python3-cffi module is not really
elegant. It introduce 'if' conditional, and most importantly, hides the abi
target when python3-cffi is not found; which make it difficult to specify it as
a test dependency (so testing of collect_stats.py can be skipped if it is not
built).

Remove the conditional and let prog_python become a disabler() when it is not
found. Disabler propagation will take care of the rest, which ensures that 1)
the abi target is always visible and 2) attempt to build abi_src and abi target
only when prog_python is found.