Increase data size of tuning_params_t->kernel_pid_max
kernel_pid_max is obtained from /proc/sys/kernel/pid_max which defaults to 32768
but can be up to 2^22, which then causes an overflow in collect_stats.py (it
tries to write the value via FFI, but actually checks the value in contrast to
C, which would just cause UB).
This should reduce the flakiness of the tests.